+++ CHANGELOG.md
@@ -10,6 +10,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Verbose detail output on exception.
+
+### Fixed
+
+- Access Token authentication.
+
 ## [1.1.0] - 2022-03-19
 
 ### Added